+++ lib/view_component/engine.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 require "rails"
+require "view_component/base"
 
 module ViewComponent
   class Engine < Rails::Engine # :nodoc:
-    config.view_component = ActiveSupport::OrderedOptions.new
-    config.view_component.preview_paths ||= []
+    config.view_component = ViewComponent::Config.default
 
     rake_tasks do
       load "view_component/rails/tasks/view_component.rake"
@@ -14,24 +14,19 @@ module ViewComponent
     initializer "view_component.set_configs" do |app|
       options = app.config.view_component
 
+      %i[generate preview_controller preview_route show_previews_source].each do |config_option|
+        options[config_option] ||= ViewComponent::Base.public_send(config_option)
+      end
+      options.instrumentation_enabled = false if options.instrumentation_enabled.nil?
       options.render_monkey_patch_enabled = true if options.render_monkey_patch_enabled.nil?
       options.show_previews = Rails.env.development? || Rails.env.test? if options.show_previews.nil?
-      options.show_previews_source ||= ViewComponent::Base.show_previews_source
       options.instrumentation_enabled = false if options.instrumentation_enabled.nil?
-      options.preview_route ||= ViewComponent::Base.preview_route
-      options.preview_controller ||= ViewComponent::Base.preview_controller
 
       if options.show_previews
+        # This is still necessary because when `config.view_component` is declared, `Rails.root` is unspecified.
         options.preview_paths << "#{Rails.root}/test/components/previews" if defined?(Rails.root) && Dir.exist?(
           "#{Rails.root}/test/components/previews"
         )
-
-        if options.preview_path.present?
-          ViewComponent::Deprecation.warn(
-            "`preview_path` will be removed in v3.0.0. Use `preview_paths` instead."
-          )
-          options.preview_paths << options.preview_path
-        end
 
         if options.show_previews_source
           require "method_source"
@@ -40,10 +35,6 @@ module ViewComponent
             MethodSource.instance_variable_set(:@lines_for_file, {})
           end
         end
-      end
-
-      ActiveSupport.on_load(:view_component) do
-        options.each { |k, v| send("#{k}=", v) if respond_to?("#{k}=") }
       end
     end
 
@@ -72,12 +63,6 @@ module ViewComponent
       end
     end
 
-    initializer "view_component.compile_config_methods" do
-      ActiveSupport.on_load(:view_component) do
-        config.compile_methods! if config.respond_to?(:compile_methods!)
-      end
-    end
-
     initializer "view_component.monkey_patch_render" do |app|
       next if Rails.version.to_f >= 6.1 || !app.config.view_component.render_monkey_patch_enabled
 
@@ -94,7 +79,7 @@ module ViewComponent
       end
     end
 
-    initializer "view_component.include_render_component" do |app|
+    initializer "view_component.include_render_component" do |_app|
       next if Rails.version.to_f >= 6.1
 
       ActiveSupport.on_load(:action_view) do
@@ -116,7 +101,7 @@ module ViewComponent
       end
     end
 
-    initializer "compiler mode" do |app|
+    initializer "compiler mode" do |_app|
       ViewComponent::Compiler.mode = if Rails.env.development? || Rails.env.test?
         ViewComponent::Compiler::DEVELOPMENT_MODE
       else
@@ -160,7 +145,7 @@ unless defined?(ViewComponent::Base)
 
   ViewComponent::Deprecation.warn(
     "This manually engine loading is deprecated and will be removed in v3.0.0. " \
-    "Remove `require \"view_component/engine\"`."
+    'Remove `require "view_component/engine"`.'
   )
 
   require "view_component"